modularize color_palette
