+++ app/models/core/style/color_palette.rb
@@ -1,0 +1,17 @@
+# == Schema Information
+#
+# Table name: color_palettes
+#
+#  id         :bigint           not null, primary key
+#  first      :string
+#  second     :string
+#  third      :string
+#  fourth     :string
+#  fifth      :string
+#  created_at :datetime         not null
+#  updated_at :datetime         not null
+#
+module Core::Style
+    class ColorPalette < ApplicationRecord
+    end
+end
